+++ crest/Assets/Crest/Crest/Scripts/LodData/Settings/SimSettingsAnimatedWaves.cs
@@ -26,7 +26,7 @@ namespace Crest
         public CollisionSources CollisionSource { get { return _collisionSource; } set { _collisionSource = value; } }
 
         [Tooltip("Maximum number of wave queries that can be performed when using ComputeShaderQueries.")]
-        [PredicatedField("_collisionSource", true, (int)CollisionSources.ComputeShaderQueries), SerializeField]
+        [Predicated("_collisionSource", true, (int)CollisionSources.ComputeShaderQueries), SerializeField, DecoratedField]
         int _maxQueryCount = QueryBase.MAX_QUERY_COUNT_DEFAULT;
         public int MaxQueryCount { get { return _maxQueryCount; } }